Further file movement refactoring.
